Rename FormSession -> Session. Also, create preview components once when initializing the FormEditProvider, rather than on every render.
