+++ packages/form-service/src/operations/submit-form.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { createForm, createFormSession } from '@atj/forms';
+import { createForm, createSession } from '@atj/forms';
 import { createTestFormService } from '../context/test';
 
 describe('submitForm', () => {
@@ -12,7 +12,7 @@ describe('submitForm', () => {
     if (!formResult.success) {
       throw new Error('form not found');
     }
-    const session = createFormSession(formResult.data);
+    const session = createSession(formResult.data);
     const result = await service.submitForm(session, 'test-form', {});
     expect(result).toEqual({
       success: true,
